3.4-django-testing/django_testing/requirements.txt
django~=4.0.2
djangorestframework~=3.13.1
psycopg2-binary

pytest~=7.0.1
django-filter~=21.1